requirements.txt
Django==1.8.2
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
bcrypt==1.1.1
billiard==3.3.0.20
boto==2.38.0
braintree==3.18.0
celery==3.1.18
cffi==1.1.2
cloudinary==1.1.2
django-boto==0.3.9
django-cities==0.4.1
django-cors-headers==1.1.0
django-filter==0.10.0
git+https://github.com/inabhi9/django-miniauth.git
django-model-utils==2.2
git+https://github.com/inabhi9/djangofuture.git
djangorestframework==3.1.3
djangorestframework-filters==0.3.0
djangorestframework-gis==0.9
djrill==1.4.0
drf-extensions==0.2.7
factory-boy==2.5.2
fake-factory==0.5.1
jsonpickle==0.9.2
kombu==3.0.26
mock==1.0.1
psycopg2==2.6
pycparser==2.13
python-dateutil==2.4.2
pytz==2015.4
raven==5.3.1
redis==2.10.3
requests==2.7.0
six==1.9.0
sqlparse==0.1.15
wsgiref==0.1.2
